+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/IndexingScrubMissing.java
@@ -213,12 +213,19 @@ public class IndexingScrubMissing extends IndexingBase {
                             index,
                             FDBRecordStoreBase.indexEntryKey(index, entry.getKey(), rec.getPrimaryKey()),
                             entry.getValue());
-                    final byte[] keyBytes = maintainer.getIndexSubspace().pack(indexEntry.getKey());
-                    return maintainer.state.transaction.get(keyBytes).thenApply(Objects::isNull);
+                    final Tuple valueKey = indexEntry.getKey();
+                    final byte[] keyBytes = maintainer.getIndexSubspace().pack(valueKey);
+                    return maintainer.state.transaction.get(keyBytes).thenApply(indexVal -> indexVal == null ? valueKey : null);
                 })
                 .collect(Collectors.toList()))
                 .thenApply(list -> {
-                    if (!list.contains(true)) {
+                    Tuple missingKey = null;
+                    for (Tuple item: list) {
+                        if (item != null) {
+                            missingKey = item;
+                        }
+                    }
+                    if (missingKey  == null) {
                         // no null index(s) = no record to index
                         return null;
                     }
@@ -227,7 +234,8 @@ public class IndexingScrubMissing extends IndexingBase {
                     if (LOGGER.isWarnEnabled() && logWarningCounter > 0) {
                         logWarningCounter --;
                         LOGGER.warn(KeyValueLogMessage.build("Scrubber: missing index entry",
-                                LogMessageKeys.KEY, rec.getPrimaryKey().toString())
+                                        LogMessageKeys.KEY, rec.getPrimaryKey().toString(),
+                                        LogMessageKeys.INDEX_KEY, missingKey.toString())
                                 .addKeysAndValues(common.indexLogMessageKeyValues())
                                 .toString());
                     }